npm i supabase:
Imports and Setup
Importing useState from React:

This hook allows you to add state management in functional components. Here, it's used to create and manage multiple state variables.
Importing supabase from a local module:

This is assumed to be an instance of the Supabase client, which is used to interact with your Supabase database.