Slice Creation Creating the cartSlice:
createSlice is used to create a slice for cart management. It takes an object with name, initialState, and reducers.
Slice Configuration:

name: Specifies the name of the slice, which is "cart" in this case.
initialState: Sets the initial state for the slice using the initialState object defined earlier.
reducers: Defines the actions and their associated state mutations.