Reducers (Actions) addToCart Action:
This action checks if the item (identified by id) is already in the cart.
If the item is present, it updates the item's quantity by incrementing it by 1.
If the item is not present, it adds the item to the cart with an initial quantity of 1.
removeFromTheCart Action:

This action removes an item from the cart based on its id.
It filters out the item with the specified id from the cart array.
incrementQuantity Action:

This action increments the quantity of a specific item in the cart.
It finds the item by id and increases its quantity by 1.
decrementQuantity Action:

This action decrements the quantity of a specific item in the cart.
It finds the item by id and decreases its quantity by 1.
clearAllCart Action:
This action clears all items from the cart, setting the cart array to an empty array.