Imports React Import:
React is imported to enable the use of JSX and React components.
Static Assets Import:

The Amazon logo image is imported from the public directory. This image will be used in the component to display the Amazon logo.
Icon Import:

The FaLock icon from the react-icons/fa6 package is imported. This icon represents a lock and is used for visual indication of security on the checkout page.
Next.js Image Component:

The Image component from next/image is imported for optimized image handling. It's used to display the Amazon logo.
Component Imports:

OrderSummary and DeliveryAddress components are imported. These components represent different parts of the checkout page and will be included in the main checkout component.
Redux Hooks and Selector Import:

The custom hook useAppSelector from the application's hooks is imported to access the Redux store.
The getCart selector from the Redux slice is imported to retrieve the cart state.