State and Logic Accessing the Cart:
The useAppSelector hook is used with the getCart selector to retrieve the cart items from the Redux store.
Calculating Total Price:

A variable totalPrice is initialized to zero.
A forEach loop iterates over each item in the cart. For each item, the total price is calculated by multiplying the item's price by its quantity and adding it to the totalPrice.