Imports React and Hooks:
React is imported to enable the use of JSX and React components.
Next.js Image Component:

The Image component from next/image is imported for optimized image handling. It helps to load and render images efficiently with built-in performance optimizations.
Ratings Component:

The Ratings component is imported. This component is used to display the product ratings.
Redux Hook:

The useAppDispatch hook is imported from @/lib/supabase/hooks/redux. This hook is likely a wrapper around useDispatch from Redux Toolkit or React-Redux, allowing components to dispatch actions to the Redux store.
Redux Action:

The addToCart action is imported from @/redux/cartSlice. This action is used to add a product to the cart.
Next.js Router:

The useRouter hook from next/navigation is imported. This hook provides methods for client-side navigation in a Next.js application.